Create timeline events from Slack pins
Adds a TimelineEvent every time a Slack pin is added and removes it when
the pin is removed.

I also decided to migrate the metadata field from
`django.contrib.postgres.fields.JSONField` to `jsonfield.JSONField` -
this supports the same features in Postgres but additionally allows the
same support in other databases - crucially, in particular, SQLite,
which means we can write tests which interact with TimelineEvents.