Add 'private' flag to incidents
Allows marking incidents to signal they contain private information.
This can be done on creation or later when editing the incident.
Once the flag is set to true it cannot be undone.

Private incidents are not omitted when listing all incidents.
The incident summary, impact, severity etc. can still be edited from
 Slack after an incident is made private.

Incident details can no longer be retrieved from the API once an
incident is made private (always returns HTTP 403).

Marking an incident as private does not automatically change the comms
channel privacy settings, this needs to be done manually.