Remove leading space from day in channel creation
In `strftime`, %e has a leading space for single digit days.  For example the 5th of the month would print as ' 5' (note the space).

Prior to this change, we were including this space in the call to channels.create, and slack was replacing the space with an extra hyphen.

This meant an incident on November 5th would show as #inc-nov--5-20-51-01 rather than the desired name of #inc-nov-5-20-51-01.

Adding a '-' in the `%e` format trims leading whitespace which corrects the problem.