Better summary and impact updates (#189)
* Better summary and impact updates
Prior to this change, the summary and impact commands would overwrite whatever was currently
set for the incident. In most cases, our understanding of these develops over time, so we'd
like to be update and refine what we've previously set.

This change introduces actions and dialogs to update the summary and impact from what was previously
set, making it easier for us to develop meaningful and thorough statements over time.