Fix broken link from incident list when UI isn't at the root
In our installation, the URLs are setup in a way that the UI part has its URLs starting with `/ui/`. When deploying the latest version, we got broken links from the new home page because it has hardcoded path, instead of using the `url` template tag to reverse by name.