DRY up signals.py
When working on a fork I needed to add another attribute to incident, which additionally required handling in signals.py, I found this required repeating blocks of code, and having extra knowledge of the codebase to be aware to update signals.py. I hoped to make an approach that allowed attributes to be added to Incident without this file being modified (unless things need customised). 

Feedback is appreciated